pyproject.toml
[project]
name = "freightlink-rop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "apache-airflow>=3.2.0",
    "boto3>=1.42.91",
    "hdbcli>=2.28.19",
    "kagglehub>=1.0.0",
    "pandas>=3.0.2",
    "pyspark==3.3.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
